Moved Upgrade interfaces and base classes into 'upgrades' package
