+++ src/main/java/com/drakmyth/minecraft/manufactory/items/upgrades/BatteryItem.java
@@ -3,7 +3,7 @@
  *  Copyright (c) 2020 Drakmyth. All rights reserved.
  */
 
-package com.drakmyth.minecraft.manufactory.items;
+package com.drakmyth.minecraft.manufactory.items.upgrades;
 
 import net.minecraft.item.Item;
 import net.minecraft.util.math.BlockPos;
