+++ src/main/java/com/drakmyth/minecraft/manufactory/items/upgrades/IPowerProvider.java
@@ -2,7 +2,7 @@
  *  SPDX-License-Identifier: LGPL-3.0-only
  *  Copyright (c) 2020 Drakmyth. All rights reserved.
  */
-package com.drakmyth.minecraft.manufactory.items;
+package com.drakmyth.minecraft.manufactory.items.upgrades;
 
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.server.ServerWorld;
