+++ src/main/java/com/drakmyth/minecraft/manufactory/items/upgrades/IPowerUpgrade.java
@@ -3,7 +3,7 @@
  *  Copyright (c) 2020 Drakmyth. All rights reserved.
  */
 
-package com.drakmyth.minecraft.manufactory.items;
+package com.drakmyth.minecraft.manufactory.items.upgrades;
 
 public interface IPowerUpgrade extends IPowerProvider {
     boolean rendersConnection();
